requirements.txt
python-decouple
pdfplumber
langchain
langchain-openai
pydantic
jinja2
streamlit
weasyprint
